fix: security & correctness fixes, responsive UI, tests/CI, dependency updates
Security & correctness (Rust backend):
- Enforce TLS certificate validation by default; add a per-server
  allow_invalid_certs opt-in (model, DB column + migration, UI toggle)
- Store auth tokens/passwords/API keys and the OpenAI key in the OS
  keychain via the keyring crate, with a plaintext-DB fallback and lazy
  migration of existing values
- Cascade-delete dependent rows on server delete; also clear sample_data
  in clear_all_data (no more orphaned rows)
- Fix validation category summary counting skipped tests as failed
- Reconstruct load-test export duration from timestamps; batch validation
  result inserts in a single transaction

Responsive UI (Angular):
- Collapse the sidenav into an overlay drawer with a toolbar hamburger
  below 960px (CDK BreakpointObserver); auto-close on navigation
- Give the drawer an opaque surface background so overlay mode no longer
  bleeds page content through
- Add breakpoints to the validation category grid; truncate the toolbar
  title; wrap the joining-property row

Tests & CI:
- Fix the stale app.spec.ts; add 10 Rust unit tests covering summary,
  percentile, JSON path resolution and field-rule validation
- Run the frontend (ng test) and Rust (cargo test) suites in CI
- Add libdbus-1-dev/pkg-config to the Linux release job for keyring

Dependencies:
- Angular 21.1 -> 21.2, Tauri 2.10 -> 2.11, vitest 4.0 -> 4.1; cargo update
- Remove duplicate @angular/material devDependency

+++ src-tauri/src/export.rs
@@ -718,7 +718,7 @@ td{padding:4px 10px;border-bottom:1px solid #e0e0e0}
                 ep_chart.x_axis().set_name("ms");
                 ep_chart.set_style(10);
                 ep_chart.set_width(460);
-                ep_chart.set_height((endpoints.len() as u32 * 28 + 120).min(380) as u32);
+                ep_chart.set_height((endpoints.len() as u32 * 28 + 120).min(380));
                 ep_chart.add_series()
                     .set_name("Avg Latency (ms)")
                     .set_categories(("Summary & Charts", ep_data_first, 5, ep_data_last, 5))
